Transcribe on GPU, with or without external LM
